refactor: Add react-icons package and update Assignment interface
The code changes include adding the "react-icons" package to the client's package.json file and updating the Assignment interface in client/src/types/assignment.ts to include the assignmentSection field. This refactor enhances the user experience by allowing the use of icons in the UserPerformanceModal component and provides additional information for assignments.