feat: Add createProblem and updateProblem methods to ProblemService
This commit adds the createProblem and updateProblem methods to the ProblemService class in the problemService.ts file. These methods make POST and PUT requests, respectively, to the corresponding API endpoints for creating and updating problems. The necessary headers and request bodies are included in the requests.

These changes enhance the functionality of the ProblemService class by providing the ability to create and update problems.