feat: Update link to edit problem in ProblemsTable component
This commit updates the link in the ProblemsTable component to navigate to the correct page for editing a problem. The previous link was pointing to `/admin/edit/problem/{problemId}`, but it has been changed to `/practice/update/{problemId}` to align with the updated routing structure.

These changes ensure that users are directed to the correct page for editing a problem in the Practice section of the application.