refactor: Add getContestUserProblemStatus method to CodeExecutionService
This commit adds the getContestUserProblemStatus method to the CodeExecutionService class in the client/src/api/codeExecutionService.ts file. The method makes a GET request to retrieve the status of a user's problem submission in a contest. It includes the necessary headers and authorization token. This change enhances the functionality of the CodeExecutionService and enables the retrieval of contest user problem status.